requirements.txt
selenium~=4.12.0
numpy~=1.24.4
scikit-fuzzy~=0.4.2
matplotlib~=3.7.3
PyYAML~=6.0.1